Testing synchronization -This is a test commit version 1
